[CORE] Backport https://github.com/bitcoin/bitcoin/commit/618d07faa2cc2f061a2e8035c3edbffc192480d7
