+++ src/miner.cpp
@@ -8,6 +8,7 @@
 #include "miner.h"
 
 #include "amount.h"
+#include "consensus/tx_verify.h"
 #include "hash.h"
 #include "main.h"
 #include "masternode-sync.h"
