+++ src/test/script_P2SH_tests.cpp
@@ -3,6 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "key.h"
+#include "consensus/tx_verify.h"
 #include "keystore.h"
 #include "main.h"
 #include "script/script.h"
