+++ src/test/sigopcount_tests.cpp
@@ -3,6 +3,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "consensus/tx_verify.h"
 #include "pubkey.h"
 #include "key.h"
 #include "script/script.h"
