+++ src/walletdb.cpp
@@ -7,6 +7,7 @@
 
 #include "walletdb.h"
 
+#include "consensus/tx_verify.h"
 #include "base58.h"
 #include "protocol.h"
 #include "serialize.h"
